refactor(modmail): split resync into snippets and panels
