+++ apps/website/src/app/dashboard/[id]/modmail/panels/page.tsx
@@ -1,5 +1,6 @@
 import { DmModeBanner } from './_components/DmModeBanner';
 import { PanelsList } from './_components/PanelsList';
+import { ResyncPanelsCard } from './_components/ResyncPanelsCard';
 import { Heading } from '@/components/common/Heading';
 import { DashboardCrumbs } from '@/components/dashboard/DashboardCrumbs';
 
@@ -13,6 +14,7 @@ export default function ModmailPanelsPage() {
 
 			<div className="grid grid-cols-1 gap-6 md:grid-cols-2 lg:grid-cols-3">
 				<DmModeBanner />
+				<ResyncPanelsCard />
 				<PanelsList />
 			</div>
 		</>
