+++ apps/website/src/app/dashboard/[id]/modmail/snippets/page.tsx
@@ -1,4 +1,4 @@
-import { ResyncCard } from './_components/ResyncCard';
+import { ResyncSnippetsCard } from './_components/ResyncSnippetsCard';
 import { SnippetsList } from './_components/SnippetsList';
 import { Heading } from '@/components/common/Heading';
 import { DashboardCrumbs } from '@/components/dashboard/DashboardCrumbs';
@@ -12,7 +12,7 @@ export default function ModmailSnippetsPage() {
 			</div>
 
 			<div className="grid grid-cols-1 items-start gap-6 md:grid-cols-2 lg:grid-cols-3">
-				<ResyncCard />
+				<ResyncSnippetsCard />
 				<SnippetsList />
 			</div>
 		</>
